requirements.txt
django
pillow
django-widget-tweaks
django_taggit
markdown
psycopg2-binary